--- back/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "deanon-back",
-  "version": "0.1.0",
-  "scripts": {
-    "start": "node ./app.js",
-    "lint": "eslint \"**/*.js\"",
-    "lint:fix": "eslint \"**/*.js\" --fix"
-  },
-  "dependencies": {
-    "express": "~4.16.1",
-    "firebase": "^6.1.1",
-    "firestore": "^1.1.6"
-  },
-  "devDependencies": {
-    "eslint": "^5.16.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.17.3",
-    "eslint-plugin-node": "^9.1.0",
-    "eslint-plugin-promise": "^4.1.1",
-    "eslint-plugin-standard": "^4.0.0"
-  }
-}
