--- back/run.bat
@@ -1,1 +1,0 @@
-node app.js
