+++ front/package.json
@@ -10,6 +10,7 @@
     "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
+    "@vue/cli": "^3.9.2",
     "axios": "^0.19.0",
     "core-js": "^2.6.5",
     "socket.io-client": "^2.2.0",
