+++ front/src/views/CreateGame.vue
@@ -48,7 +48,7 @@ export default {
   },
   methods: {
     createGame () {
-      axios.post(`${this.$store.getters.getAPI_URL}create`, this.gameSettings)
+      axios.post(`${this.$store.getters.getAPI_URL}games/create`, this.gameSettings)
         .then((response) => {
           if (!response.data.id) {
             this.errorList.push('Не было получено id игры :(')
